src/requirements.txt
mysql-connector-python
kafka-python
pandas
pytest
